Updating C++ Source codes

+++ Jim_v7/Source_Code/Tiff_Channel_Splitter/Orientate_Images.cpp
@@ -3,9 +3,9 @@
 
 using namespace std;
 
-void vertFlipImage(std::vector< std::vector<float>>& imageio) {
+void vertFlipImage(std::vector< std::vector<uint16_t>>& imageio) {
 
-	vector<vector<float>> imageout(imageio.size(),vector<float>(imageio[0].size(),0.0));
+	vector<vector<uint16_t>> imageout(imageio.size(),vector<uint16_t>(imageio[0].size(),0.0));
 	uint32_t imageHeight = imageio[0].size();
 
 	for (int i = 0; i < imageio.size(); i++)for (int j = 0; j < imageio[0].size(); j++)imageout[i][imageHeight - j - 1] = imageio[i][j];
@@ -13,9 +13,9 @@ void vertFlipImage(std::vector< std::vector<float>>& imageio) {
 
 }
 
-void horzFlipImage(std::vector< std::vector<float>>& imageio) {
+void horzFlipImage(std::vector< std::vector<uint16_t>>& imageio) {
 
-	vector<vector<float>> imageout(imageio.size(), vector<float>(imageio[0].size(), 0.0));
+	vector<vector<uint16_t>> imageout(imageio.size(), vector<uint16_t>(imageio[0].size(), 0.0));
 	uint32_t imageWidth = imageio.size();
 
 	for (int i = 0; i < imageio.size(); i++)for (int j = 0; j < imageio[0].size(); j++)imageout[imageWidth-1-i][j] = imageio[i][j];
@@ -23,22 +23,22 @@ void horzFlipImage(std::vector< std::vector<float>>& imageio) {
 
 }
 
-void rotateImage(std::vector< std::vector<float>>& imageio,int angle) {
+void rotateImage(std::vector< std::vector<uint16_t>>& imageio,int angle) {
 	uint32_t imageWidth = imageio.size();
 	uint32_t imageHeight = imageio[0].size();
 
 	if (angle == 180) {
-		vector<vector<float>> imageout(imageio.size(), vector<float>(imageio[0].size(), 0.0));
+		vector<vector<uint16_t>> imageout(imageio.size(), vector<uint16_t>(imageio[0].size(), 0.0));
 		for (int i = 0; i < imageio.size(); i++)for (int j = 0; j < imageio[0].size(); j++)imageout[imageWidth - 1 - i][imageHeight - 1 - j] = imageio[i][j];
 		imageio = imageout;
 	}
 	else if (angle == 90) {
-		vector<vector<float>> imageout(imageio[0].size(), vector<float>(imageio.size(), 0.0));
+		vector<vector<uint16_t>> imageout(imageio[0].size(), vector<uint16_t>(imageio.size(), 0.0));
 		for (int i = 0; i < imageio.size(); i++)for (int j = 0; j < imageio[0].size(); j++)imageout[imageHeight - 1 - j][i] = imageio[i][j];
 		imageio = imageout;
 	}
 	else if (angle == 270) {
-		vector<vector<float>> imageout(imageio[0].size(), vector<float>(imageio.size(), 0.0));
+		vector<vector<uint16_t>> imageout(imageio[0].size(), vector<uint16_t>(imageio.size(), 0.0));
 		for (int i = 0; i < imageio.size(); i++)for (int j = 0; j < imageio[0].size(); j++)imageout[j][imageWidth - 1 - i] = imageio[i][j];
 		imageio = imageout;
 	}
